+++ lib/urn.rb
@@ -23,11 +23,30 @@ class URN
     true
   end
 
+  def nid
+    _scheme, nid, _nss = parse
+
+    nid
+  end
+
+  def nss
+    _scheme, _nid, nss = parse
+
+    nss
+  end
+
   def normalize
-    _scheme, nid, nss = urn.split(':', 3)
+    _scheme, nid, nss = parse
+
     normalized_nid = nid.downcase
     normalized_nss = nss.gsub(/%([0-9a-f]{2})/i, &:downcase)
 
     "urn:#{normalized_nid}:#{normalized_nss}"
+  end
+
+  private
+
+  def parse
+    urn.split(':', 3)
   end
 end